Fix handling of suppress modularization
Partially fixes wixtoolset/issues#5944
