+++ src/test/WixToolsetTest.CoreIntegration/TestData/SuppressModularization/Module.en-us.wxl
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+This file contains the declaration of all the localizable strings.
+-->
+<WixLocalization xmlns="http://wixtoolset.org/schemas/v4/wxl" Culture="en-US">
+
+  <String Id="Manufacturer">Example Company</String>
+
+</WixLocalization>
