+++ src/test/WixToolsetTest.CoreIntegration/TestData/SuppressModularization/data/test.txt
@@ -1,0 +1,1 @@
+This is test.txt.
